update get payment method tests
